[hla_plugin] receive cbrn data
Conflicts:

	src/libraries/hla_plugin/HlaPlugin.cpp
	src/libraries/hla_plugin/SimulationFacade.cpp
	src/libraries/hla_plugin/SimulationFacade.h
	src/libraries/hla_plugin/UnitTeleporter.cpp

+++ src/libraries/hla_plugin/RawDataHazardContourGroup.h
@@ -83,7 +83,7 @@ private:
     std::string identifier_;
     std::auto_ptr< AttributesUpdater > attributes_;
     uint64_t time_;
-    uint32_t material_;
+    uint16_t material_;
     uint8_t hazardType_;
     std::vector< RawDataHazardContour > contours_;
     //@}
